requirements.txt
bittensor==9.1.0
bittensor-wallet==3.0.4
atomics==1.0.3
starlette==0.37.2
pydantic==2.9.2
pytest==8.3.3
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2
py-solc-x==2.0.3
solc_ast_parser>=1.3
hypothesis==6.124.9
solidity-vulnerabilities-db<1
websocket-client==1.8.0
async-substrate-interface>=1.0.0
unique-playgrounds-py==0.0.2
solidity_audit_lib @ git+https://github.com/ReinforcedAIAudits/lib.git
